+++ attribution/business/attribution_charge_new.py
@@ -32,4 +32,4 @@ def find_attribution_charge_new_by_learning_unit_year(learning_unit_year):
     return AttributionChargeNew.objects.filter(
         learning_component_year__in=[learning_unit_component.learning_component_year
             for learning_unit_component in learning_unit_components ]
-    )
+    ).distinct('attribution__tutor').select_related('attribution__tutor__person')
