+++ base/templates/learning_unit/attributions.html
@@ -58,9 +58,9 @@
                                 </td>
                                 <td></td>
                                 <td>{{ attribution_charge_new.attribution.start_year | default_if_none:'' }}</td>
-                                <td>{{ attribution_charge_new.attribution.duration | default_if_none:'' }}</td>
-                                <td>{{ attribution_charge_new.attribution.volume_lecturing  | default_if_none:'' }}</td>
-                                <td>{{ attribution_charge_new.attribution.volume_practical  | default_if_none:'' }}</td>
+                                <td>{{ attribution_charge_new.duration | default_if_none:'' }}</td>
+                                <td>{{ attribution_charge_new.volume_lecturing  | default_if_none:'' }}</td>
+                                <td>{{ attribution_charge_new.volume_practical  | default_if_none:'' }}</td>
                             </tr>
                         {% endfor %}
                     </tbody>
